Revert "Revert "refactored few type objects""
This reverts commit 926a98677790afae9af543a188f2f362485b7bc4.
