+++ pkg/pipeline/SystemWorkflowExecutor_test.go
@@ -8,6 +8,7 @@ import (
 	bean2 "github.com/devtron-labs/devtron/api/bean"
 	"github.com/devtron-labs/devtron/internal/util"
 	"github.com/devtron-labs/devtron/pkg/pipeline/bean"
+	"github.com/devtron-labs/devtron/pkg/pipeline/types"
 	"github.com/stretchr/testify/assert"
 	v1 "k8s.io/api/batch/v1"
 	v12 "k8s.io/api/core/v1"
@@ -24,7 +25,7 @@ func TestSystemWorkflowExecute(t *testing.T) {
 	t.SkipNow()
 	logger, loggerErr := util.NewSugardLogger()
 	assert.Nil(t, loggerErr)
-	cdConfig, err := GetCiCdConfig()
+	cdConfig, err := types.GetCiCdConfig()
 	assert.Nil(t, err)
 	runtimeConfig, err := client.GetRuntimeConfig()
 	assert.Nil(t, err)
