+++ pkg/pipeline/WebhookService.go
@@ -29,6 +29,7 @@ import (
 	"github.com/devtron-labs/devtron/internal/sql/repository/pipelineConfig"
 	util2 "github.com/devtron-labs/devtron/internal/util"
 	"github.com/devtron-labs/devtron/pkg/app"
+	types2 "github.com/devtron-labs/devtron/pkg/pipeline/types"
 	"github.com/devtron-labs/devtron/pkg/sql"
 	"github.com/devtron-labs/devtron/util/event"
 	"github.com/go-pg/pg"
@@ -61,7 +62,7 @@ type WebhookService interface {
 
 type WebhookServiceImpl struct {
 	ciArtifactRepository repository.CiArtifactRepository
-	ciConfig             *CiConfig
+	ciConfig             *types2.CiConfig
 	logger               *zap.SugaredLogger
 	ciPipelineRepository pipelineConfig.CiPipelineRepository
 	ciWorkflowRepository pipelineConfig.CiWorkflowRepository
@@ -94,7 +95,7 @@ func NewWebhookServiceImpl(
 		ciHandler:            ciHandler,
 		customTagService:     customTagService,
 	}
-	config, err := GetCiConfig()
+	config, err := types2.GetCiConfig()
 	if err != nil {
 		return nil
 	}
