+++ pkg/pipeline/WorkflowService.go
@@ -31,6 +31,7 @@ import (
 	"github.com/devtron-labs/devtron/pkg/cluster/repository"
 	k8s2 "github.com/devtron-labs/devtron/pkg/k8s"
 	bean3 "github.com/devtron-labs/devtron/pkg/pipeline/bean"
+	"github.com/devtron-labs/devtron/pkg/pipeline/types"
 	"go.uber.org/zap"
 	v12 "k8s.io/api/core/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -42,7 +43,7 @@ import (
 // TODO: move isCi/isJob to workflowRequest
 
 type WorkflowService interface {
-	SubmitWorkflow(workflowRequest *WorkflowRequest) (*unstructured.UnstructuredList, error)
+	SubmitWorkflow(workflowRequest *types.WorkflowRequest) (*unstructured.UnstructuredList, error)
 	//DeleteWorkflow(wfName string, namespace string) error
 	GetWorkflow(name string, namespace string, isExt bool, environment *repository.Environment) (*v1alpha1.Workflow, error)
 	//ListAllWorkflows(namespace string) (*v1alpha1.WorkflowList, error)
@@ -53,7 +54,7 @@ type WorkflowService interface {
 type WorkflowServiceImpl struct {
 	Logger                 *zap.SugaredLogger
 	config                 *rest.Config
-	ciCdConfig             *CiCdConfig
+	ciCdConfig             *types.CiCdConfig
 	appService             app.AppService
 	envRepository          repository.EnvironmentRepository
 	globalCMCSService      GlobalCMCSService
@@ -65,7 +66,7 @@ type WorkflowServiceImpl struct {
 
 // TODO: Move to bean
 
-func NewWorkflowServiceImpl(Logger *zap.SugaredLogger, envRepository repository.EnvironmentRepository, ciCdConfig *CiCdConfig,
+func NewWorkflowServiceImpl(Logger *zap.SugaredLogger, envRepository repository.EnvironmentRepository, ciCdConfig *types.CiCdConfig,
 	appService app.AppService, globalCMCSService GlobalCMCSService, argoWorkflowExecutor ArgoWorkflowExecutor,
 	k8sUtil *k8s.K8sUtil,
 	systemWorkflowExecutor SystemWorkflowExecutor, k8sCommonService k8s2.K8sCommonService) (*WorkflowServiceImpl, error) {
@@ -89,20 +90,13 @@ func NewWorkflowServiceImpl(Logger *zap.SugaredLogger, envRepository repository.
 }
 
 const (
-	BLOB_STORAGE_AZURE             = "AZURE"
-	BLOB_STORAGE_S3                = "S3"
-	BLOB_STORAGE_GCP               = "GCP"
-	BLOB_STORAGE_MINIO             = "MINIO"
 	CI_NODE_SELECTOR_APP_LABEL_KEY = "devtron.ai/node-selector"
-	CI_NODE_PVC_ALL_ENV            = "devtron.ai/ci-pvc-all"
-	CI_NODE_PVC_PIPELINE_PREFIX    = "devtron.ai/ci-pvc"
-	PRE                            = "PRE"
-	POST                           = "POST"
-	preCdStage                     = "preCD"
-	postCdStage                    = "postCD"
+
+	preCdStage  = "preCD"
+	postCdStage = "postCD"
 )
 
-func (impl *WorkflowServiceImpl) SubmitWorkflow(workflowRequest *WorkflowRequest) (*unstructured.UnstructuredList, error) {
+func (impl *WorkflowServiceImpl) SubmitWorkflow(workflowRequest *types.WorkflowRequest) (*unstructured.UnstructuredList, error) {
 	workflowTemplate, err := impl.createWorkflowTemplate(workflowRequest)
 	if err != nil {
 		return nil, err
@@ -115,7 +109,7 @@ func (impl *WorkflowServiceImpl) SubmitWorkflow(workflowRequest *WorkflowRequest
 	return createdWf, err
 }
 
-func (impl *WorkflowServiceImpl) createWorkflowTemplate(workflowRequest *WorkflowRequest) (bean3.WorkflowTemplate, error) {
+func (impl *WorkflowServiceImpl) createWorkflowTemplate(workflowRequest *types.WorkflowRequest) (bean3.WorkflowTemplate, error) {
 	workflowJson, err := workflowRequest.GetWorkflowJson(impl.ciCdConfig)
 	if err != nil {
 		impl.Logger.Errorw("error occurred while getting workflow json", "err", err)
@@ -164,7 +158,7 @@ func (impl *WorkflowServiceImpl) createWorkflowTemplate(workflowRequest *Workflo
 	return workflowTemplate, nil
 }
 
-func (impl *WorkflowServiceImpl) getClusterConfig(workflowRequest *WorkflowRequest) (*rest.Config, error) {
+func (impl *WorkflowServiceImpl) getClusterConfig(workflowRequest *types.WorkflowRequest) (*rest.Config, error) {
 	env := workflowRequest.Env
 	if workflowRequest.IsExtRun {
 		configMap := env.Cluster.Config
@@ -186,7 +180,7 @@ func (impl *WorkflowServiceImpl) getClusterConfig(workflowRequest *WorkflowReque
 
 }
 
-func (impl *WorkflowServiceImpl) appendGlobalCMCS(workflowRequest *WorkflowRequest) ([]bean.ConfigSecretMap, []bean.ConfigSecretMap, error) {
+func (impl *WorkflowServiceImpl) appendGlobalCMCS(workflowRequest *types.WorkflowRequest) ([]bean.ConfigSecretMap, []bean.ConfigSecretMap, error) {
 	var workflowConfigMaps []bean.ConfigSecretMap
 	var workflowSecrets []bean.ConfigSecretMap
 	if !workflowRequest.IsExtRun {
@@ -208,7 +202,7 @@ func (impl *WorkflowServiceImpl) appendGlobalCMCS(workflowRequest *WorkflowReque
 	return workflowConfigMaps, workflowSecrets, nil
 }
 
-func (impl *WorkflowServiceImpl) addExistingCmCsInWorkflow(workflowRequest *WorkflowRequest, workflowConfigMaps []bean.ConfigSecretMap, workflowSecrets []bean.ConfigSecretMap) ([]bean.ConfigSecretMap, []bean.ConfigSecretMap, error) {
+func (impl *WorkflowServiceImpl) addExistingCmCsInWorkflow(workflowRequest *types.WorkflowRequest, workflowConfigMaps []bean.ConfigSecretMap, workflowSecrets []bean.ConfigSecretMap) ([]bean.ConfigSecretMap, []bean.ConfigSecretMap, error) {
 
 	pipelineLevelConfigMaps, pipelineLevelSecrets, err := workflowRequest.GetConfiguredCmCs()
 	if err != nil {
@@ -246,7 +240,7 @@ func (impl *WorkflowServiceImpl) addExistingCmCsInWorkflow(workflowRequest *Work
 	return workflowConfigMaps, workflowSecrets, nil
 }
 
-func (impl *WorkflowServiceImpl) updateBlobStorageConfig(workflowRequest *WorkflowRequest, workflowTemplate *bean3.WorkflowTemplate) {
+func (impl *WorkflowServiceImpl) updateBlobStorageConfig(workflowRequest *types.WorkflowRequest, workflowTemplate *bean3.WorkflowTemplate) {
 	workflowTemplate.BlobStorageConfigured = workflowRequest.BlobStorageConfigured && (workflowRequest.CheckBlobStorageConfig(impl.ciCdConfig) || !workflowRequest.IsExtRun)
 	workflowTemplate.BlobStorageS3Config = workflowRequest.BlobStorageS3Config
 	workflowTemplate.AzureBlobConfig = workflowRequest.AzureBlobConfig
@@ -254,7 +248,7 @@ func (impl *WorkflowServiceImpl) updateBlobStorageConfig(workflowRequest *Workfl
 	workflowTemplate.CloudStorageKey = workflowRequest.BlobStorageLogsKey
 }
 
-func (impl *WorkflowServiceImpl) getAppLabelNodeSelector(workflowRequest *WorkflowRequest) map[string]string {
+func (impl *WorkflowServiceImpl) getAppLabelNodeSelector(workflowRequest *types.WorkflowRequest) map[string]string {
 	// node selector
 	if val, ok := workflowRequest.AppLabels[CI_NODE_SELECTOR_APP_LABEL_KEY]; ok && !(workflowRequest.CheckForJob() && workflowRequest.IsExtRun) {
 		var nodeSelectors map[string]string
